[refactor] re-pacakage & check Deprecated class
