+++ src/main/kotlin/com/meokq/api/file/annotations/ExplainSaveImage.kt
@@ -1,4 +1,4 @@
-package com.meokq.api.image.annotations
+package com.meokq.api.file.annotations
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
