+++ src/main/kotlin/com/meokq/api/file/repository/ImageRepository.kt
@@ -1,6 +1,6 @@
-package com.meokq.api.image.repository
+package com.meokq.api.file.repository
 
-import com.meokq.api.image.model.Image
+import com.meokq.api.file.model.Image
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ImageRepository : JpaRepository<Image, String> {
